create BroadcastContainer component
